removed some code duplication in tests
