+++ tests/test_htmy.py
@@ -15,24 +15,13 @@ from .data import (
     user_list_json,
     users,
 )
+from .errors import RenderedError
 from .htmy_components import HelloWorld, Profile, RequestProcessors, UserList
 
 billy_html_header = "<h1 >Billy Shears (active=True)</h1>"
 billy_html_paragraph = "<p >Billy Shears (active=True)</p>"
 billy_html_span = "<span >Billy Shears (active=True)</span>"
 user_list_html = "<ul >\n<li >Billy Shears (active=True)</li>\n<li >Lucy (active=True)</li>\n</ul>"
-
-
-class RenderedError(Exception):
-    def __init__(self, data: dict[str, Any], *, response: Response) -> None:
-        super().__init__("Data validation failed.")
-
-        # Pattern for setting the response status code for error rendering responses.
-        response.status_code = 456
-
-        # Pattern to make the data available in rendering contexts. Not used in tests.
-        for key, value in data.items():
-            setattr(self, key, value)
 
 
 @pytest.fixture
