+++ tests/test_jinja.py
@@ -20,18 +20,7 @@ from .data import (
     user_list_json,
     users,
 )
-
-
-class RenderedError(Exception):
-    def __init__(self, data: dict[str, Any], *, response: Response) -> None:
-        super().__init__("Data validation failed.")
-
-        # Pattern for setting the response status code for error rendering responses.
-        response.status_code = 456
-
-        # Pattern to make the data available in Jinja rendering contexts. Not used in tests.
-        for key, value in data.items():
-            setattr(self, key, value)
+from .errors import RenderedError
 
 
 @pytest.fixture
